chore: update dependencies and refactor prisma client imports
- Updated various dependencies in package.json to their latest versions.
- Changed the output path for the Prisma client in schema.prisma.
- Refactored imports in multiple files to use the generated Prisma client instead of the original path.
- Updated the import statements for Radix UI components to use the new package structure.
