+++ .env.example
@@ -1,0 +1,7 @@
+# Supabase
+NEXT_PUBLIC_SUPABASE_URL=""
+NEXT_PUBLIC_SUPABASE_PUBLISHABLE_OR_ANON_KEY=""
+
+# Prisma
+DATABASE_URL=""
+DIRECT_URL=""
