+++ package.json
@@ -10,15 +10,11 @@
     "format": "prettier . --write"
   },
   "dependencies": {
-    "@hookform/resolvers": "^5.2.1",
-    "@prisma/client": "^6.16.2",
-    "@radix-ui/react-label": "^2.1.7",
-    "@radix-ui/react-separator": "^1.1.7",
-    "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-use-controllable-state": "^1.2.2",
+    "@hookform/resolvers": "^5.2.2",
+    "@prisma/client": "^6.19.0",
     "@supabase/ssr": "^0.7.0",
-    "@supabase/supabase-js": "^2.57.2",
-    "@tanstack/react-query": "^5.87.1",
+    "@supabase/supabase-js": "^2.81.1",
+    "@tanstack/react-query": "^5.90.8",
     "@tanstack/react-table": "^8.21.3",
     "@vercel/analytics": "^1.5.0",
     "@vercel/speed-insights": "^1.2.0",
@@ -28,38 +24,38 @@
     "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
-    "lucide-react": "^0.544.0",
-    "next": "^16.0.0",
+    "lucide-react": "^0.553.0",
+    "next": "^16.0.1",
     "next-themes": "^0.4.6",
-    "nuqs": "^2.6.0",
+    "nuqs": "^2.7.3",
     "radix-ui": "^1.4.3",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
-    "react-hook-form": "^7.62.0",
-    "sharp": "^0.34.3",
+    "react-hook-form": "^7.66.0",
+    "sharp": "^0.34.5",
     "slugify": "^1.6.6",
     "sonner": "^2.0.7",
-    "tailwind-merge": "^3.3.1",
+    "tailwind-merge": "^3.4.0",
     "vaul": "^1.1.2",
-    "zod": "^4.1.5",
+    "zod": "^4.1.12",
     "zustand": "^5.0.8"
   },
   "devDependencies": {
-    "@biomejs/biome": "2.2.2",
+    "@biomejs/biome": "2.3.5",
     "@ianvs/prettier-plugin-sort-imports": "^4.7.0",
-    "@tailwindcss/postcss": "^4",
-    "@tanstack/react-query-devtools": "^5.87.1",
-    "@types/node": "^24.3.1",
-    "@types/react": "^19.2.2",
-    "@types/react-dom": "^19.2.2",
+    "@tailwindcss/postcss": "^4.1.17",
+    "@tanstack/react-query-devtools": "^5.90.2",
+    "@types/node": "^24.10.1",
+    "@types/react": "^19.2.3",
+    "@types/react-dom": "^19.2.3",
     "husky": "^9.1.7",
-    "lint-staged": "^16.1.6",
+    "lint-staged": "^16.2.6",
     "prettier": "^3.6.2",
-    "prettier-plugin-tailwindcss": "^0.6.14",
-    "prisma": "^6.16.2",
-    "tailwindcss": "^4.1.13",
-    "tw-animate-css": "^1.3.8",
-    "typescript": "^5",
-    "ultracite": "5.3.3"
+    "prettier-plugin-tailwindcss": "^0.7.1",
+    "prisma": "^6.19.0",
+    "tailwindcss": "^4.1.17",
+    "tw-animate-css": "^1.4.0",
+    "typescript": "^5.9.3",
+    "ultracite": "6.3.2"
   }
 }
