+++ prisma/schema.prisma
@@ -1,6 +1,6 @@
 generator client {
   provider = "prisma-client"
-  output   = "../src/lib/prisma"
+  output   = "../src/lib/generated/prisma"
 }
 
 datasource db {
