+++ src/components/ui/file-upload.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import { Slot } from "@radix-ui/react-slot";
+import { Slot as SlotPrimitive } from "radix-ui";
 import { FileArchiveIcon, FileAudioIcon, FileCodeIcon, FileCogIcon, FileIcon, FileTextIcon, FileVideoIcon } from "lucide-react";
 
 import { cn } from "@/lib/utils";
@@ -577,7 +577,7 @@ function FileUploadRoot(props: FileUploadRootProps) {
 		[dropzoneId, inputId, listId, labelId, dir, disabled, urlCache]
 	);
 
-	const RootPrimitive = asChild ? Slot : "div";
+	const RootPrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	return (
 		<StoreContext.Provider value={store}>
@@ -772,7 +772,7 @@ function FileUploadDropzone(props: FileUploadDropzoneProps) {
 		[context.inputRef, onKeyDownProp]
 	);
 
-	const DropzonePrimitive = asChild ? Slot : "div";
+	const DropzonePrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	return (
 		<DropzonePrimitive
@@ -821,7 +821,7 @@ function FileUploadTrigger(props: FileUploadTriggerProps) {
 		[context.inputRef, onClickProp]
 	);
 
-	const TriggerPrimitive = asChild ? Slot : "button";
+	const TriggerPrimitive = asChild ? SlotPrimitive.Slot : "button";
 
 	return (
 		<TriggerPrimitive
@@ -851,7 +851,7 @@ function FileUploadList(props: FileUploadListProps) {
 
 	if (!shouldRender) return null;
 
-	const ListPrimitive = asChild ? Slot : "div";
+	const ListPrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	return (
 		<ListPrimitive
@@ -934,7 +934,7 @@ function FileUploadItem(props: FileUploadItemProps) {
 				? "Upload complete"
 				: "Ready to upload";
 
-	const ItemPrimitive = asChild ? Slot : "div";
+	const ItemPrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	return (
 		<FileUploadItemContext.Provider value={itemContext}>
@@ -1028,7 +1028,7 @@ function FileUploadItemPreview(props: FileUploadItemPreviewProps) {
 
 	if (!itemContext.fileState) return null;
 
-	const ItemPreviewPrimitive = asChild ? Slot : "div";
+	const ItemPreviewPrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	return (
 		<ItemPreviewPrimitive
@@ -1056,7 +1056,7 @@ function FileUploadItemMetadata(props: FileUploadItemMetadataProps) {
 
 	if (!itemContext.fileState) return null;
 
-	const ItemMetadataPrimitive = asChild ? Slot : "div";
+	const ItemMetadataPrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	return (
 		<ItemMetadataPrimitive
@@ -1107,7 +1107,7 @@ function FileUploadItemProgress(props: FileUploadItemProgressProps) {
 
 	if (!shouldRender) return null;
 
-	const ItemProgressPrimitive = asChild ? Slot : "div";
+	const ItemProgressPrimitive = asChild ? SlotPrimitive.Slot : "div";
 
 	switch (variant) {
 		case "circular": {
@@ -1225,7 +1225,7 @@ function FileUploadItemDelete(props: FileUploadItemDeleteProps) {
 
 	if (!itemContext.fileState) return null;
 
-	const ItemDeletePrimitive = asChild ? Slot : "button";
+	const ItemDeletePrimitive = asChild ? SlotPrimitive.Slot : "button";
 
 	return (
 		<ItemDeletePrimitive
@@ -1268,7 +1268,7 @@ function FileUploadClear(props: FileUploadClearProps) {
 
 	if (!shouldRender) return null;
 
-	const ClearPrimitive = asChild ? Slot : "button";
+	const ClearPrimitive = asChild ? SlotPrimitive.Slot : "button";
 
 	return (
 		<ClearPrimitive
