+++ src/lib/db.ts
@@ -1,4 +1,4 @@
-import { PrismaClient } from "@/lib/prisma/client";
+import { PrismaClient } from "@/lib/generated/prisma/client";
 
 const globalForPrisma = global as unknown as { prisma: PrismaClient };
